+++ safe/gui/tools/test/test_multi_exposure.py
@@ -8,7 +8,8 @@ from qgis.gui import QgsMapCanvasLayer
 from safe.definitions.exposure import exposure_road, exposure_population
 from safe.gui.tools.multi_exposure_dialog import (
     MultiExposureDialog)
-from safe.test.utilities import load_test_vector_layer, qgis_app
+from safe.test.utilities import load_test_vector_layer
+from safe.test.qgis_app import qgis_app
 
 __copyright__ = "Copyright 2017, The InaSAFE Project"
 __license__ = "GPL version 3"
